+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Viewer: optional gesture to show previous/next item
 - Albums / Countries / Tags: live title filter
 - option to hide confirmation message after moving items to the bin
+- Collection / Info: edit description via Exif / IPTC / XMP
 
 ### Changed